+++ src/main/java/site/doto/domain/character/entity/CharacterType.java
@@ -1,4 +1,4 @@
-package site.doto.domain.item_type.entity;
+package site.doto.domain.character.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -12,13 +12,17 @@ import javax.persistence.*;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class ItemType {
+public class CharacterType {
     @Id
-    @Column(name = "item_type_id")
+    @Column(name = "character_type_id")
     private Long id;
 
     private String name;
 
     private String img;
+
+    private Integer level;
+
+    private Integer species;
 
 }
